+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -176,6 +176,18 @@ pub(crate) fn migrate_eslint_any_rule(
                 .get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
+        "@typescript-eslint/no-throw-literal" => {
+            if !options.include_inspired {
+                results.has_inspired_rules = true;
+                return false;
+            }
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group.use_throw_only_error.get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
         "@typescript-eslint/no-unnecessary-type-constraint" => {
             let group = rules.complexity.get_or_insert_with(Default::default);
             let rule = group
@@ -991,6 +1003,18 @@ pub(crate) fn migrate_eslint_any_rule(
             let rule = group.no_unreachable_super.get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
+        "no-throw-literal" => {
+            if !options.include_inspired {
+                results.has_inspired_rules = true;
+                return false;
+            }
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group.use_throw_only_error.get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
         "no-undef" => {
             let group = rules.correctness.get_or_insert_with(Default::default);
             let rule = group
@@ -1110,6 +1134,18 @@ pub(crate) fn migrate_eslint_any_rule(
             let rule = group
                 .use_single_var_declarator
                 .get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
+        "only-throw-error" => {
+            if !options.include_inspired {
+                results.has_inspired_rules = true;
+                return false;
+            }
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group.use_throw_only_error.get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
         "operator-assignment" => {